+++ templates/index.html
@@ -92,12 +92,7 @@
 
 
 <script src="https://cdnjs.cloudflare.com/ajax/libs/Sortable/1.14.0/Sortable.min.js"></script>
-
 <script src = "{{ url_for('static', filename="js/themer.js")}}"></script>
-<script type = "module" src = "{{ url_for('static', filename="js/main.js")}}"></script>
+<script src = "{{ url_for('static', filename="js/main.js")}}"></script>
 <script src = "{{ url_for('static', filename="js/searchBarAnimation.js")}}"></script>
-<script type = "module" src = "{{ url_for('static', filename="js/convertor.js")}}"></script>
-
-
-
 </html>
